pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "irbstudio"
version = "0.1.0"
description = "An AIRB Scenario & Impact Analysis Engine"
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "pandas",
    "numpy",
    "pyyaml",
    "pydantic",
    "scipy",
    "plotly",
    "scikit-learn",
    "statsmodels",
    'jinja2',
    'matplotlib',
]

[project.optional-dependencies]
dev = [
    "pytest",
    "black",
    "isort",
]
